fixed errors and startder Text viewer refactor
