+++ src/components/core/organisms/FileItem/FileItem.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
-import FileItemStyled from '../../../styled/molecules/fileItemStyled'
+import FileItemStyled from '../../../styled/organisms/fileItemStyled'
 import colorScheme from '../../../../misc/colorScheme'
 import Icon from '../../atoms/Icon/Icon'
-import TextViewer from '../TextViewer/TextViewer'
+import TextViewer from '../../molecules/TextViewer/TextViewer'
 import { removeFileAction } from '../../../../redux/actions/addFileAction'
 import { connect } from 'react-redux'
 import { fileItemTextGenerator } from '../../../../misc/fileItemText'
