+++ src/components/core/pages/HomePage/HomePage.js
@@ -5,7 +5,7 @@ import Text from '../../atoms/Text/Text'
 import Input from '../../atoms/Input/Input'
 import DragAndDrop from '../../atoms/DragAndDrop/DragAndDrop'
 import List from '../../organisms/List/List'
-import FileItem from '../../molecules/FileItem/FileItem'
+import FileItem from '../../organisms/FileItem/FileItem'
 import ProfItem from '../../organisms/ProfItem/ProfItem'
 import CheckboxMessage from '../../molecules/CheckboxMessage/CheckboxMessage'
 import { concernTrigger } from '../../../../redux/actions/concernAction'
